+++ shared/src/commonMain/kotlin/com.cheesecake.mafia/data/GameAction.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class GameAction(
-    val dayIndex: Int = 0,
+    val dayIndex: Byte = 0,
     val actionType: GameActionType,
 )
